# BPF proxy startup fix