## Description

From O(n^2) towards O(n)

before:

```
BenchmarkStartupSync/Services_10_Endpoints_1-12         	   63256	     18718 ns/op	    8357 B/op	     236 allocs/op
BenchmarkStartupSync/Services_10_Endpoints_10-12        	   22261	     54462 ns/op	   18195 B/op	     545 allocs/op
BenchmarkStartupSync/Services_100_Endpoints_1-12        	    1756	    737219 ns/op	  237768 B/op	   11326 allocs/op
BenchmarkStartupSync/Services_100_Endpoints_10-12       	    1060	   1271850 ns/op	  336121 B/op	   14413 allocs/op
BenchmarkStartupSync/Services_1000_Endpoints_10-12      	      18	  63691920 ns/op	17817078 B/op	 1040328 allocs/op
BenchmarkStartupSync/Services_1000_Endpoints_100-12     	      12	  92424672 ns/op	26530761 B/op	 1319624 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_1-12      	       1	6289452353 ns/op	1605454416 B/op	100324524 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_10-12     	       1	6325168548 ns/op	1607597344 B/op	100173814 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_100-12    	       1	7681977134 ns/op	1720611008 B/op	104589689 allocs/op
```

after:

```
BenchmarkStartupSync/Services_10_Endpoints_1-12         	   74241	     16182 ns/op	    9576 B/op	     168 allocs/op
BenchmarkStartupSync/Services_10_Endpoints_10-12        	   23649	     47887 ns/op	   19417 B/op	     478 allocs/op
BenchmarkStartupSync/Services_100_Endpoints_1-12        	    7606	    152907 ns/op	  102965 B/op	    1624 allocs/op
BenchmarkStartupSync/Services_100_Endpoints_10-12       	    2353	    577624 ns/op	  202049 B/op	    4724 allocs/op
BenchmarkStartupSync/Services_1000_Endpoints_10-12      	     234	   5063275 ns/op	 2443199 B/op	   47079 allocs/op
BenchmarkStartupSync/Services_1000_Endpoints_100-12     	      36	  31615645 ns/op	11065896 B/op	  321077 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_1-12      	      64	  18008992 ns/op	11967414 B/op	  160594 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_10-12     	      22	  55656731 ns/op	21884698 B/op	  470594 allocs/op
BenchmarkStartupSync/Services_10000_Endpoints_100-12    	       3	 354862182 ns/op	108117866 B/op	 3210587 allocs/op
```

<!-- A few sentences describing the overall goals of the pull request's commits.
Please include
- the type of fix - (e.g. bug fix, new feature, documentation)
- some details on _why_ this PR should be merged
- the details of the testing you've done on it (both manual and automated)
- which components are affected by this PR
-->

## Todos
- [ ] Unit tests (full coverage)
- [ ] Integration tests (delete as appropriate) In plan/Not needed/Done
- [ ] Documentation
- [ ] Backport
- [ ] Release note

## Release Note
<!-- Writing a release note:
- By default, no release note action is required.
- If you're unsure whether or not your PR needs a note, ask your reviewer for guidance.
- If this PR requires a release note, update the block below to include a concise note describing
  the change and any important impacts this PR may have.
-->

```release-note
In BPF mode, Felix now uses a more efficient algorithm to resync the Kubernetes sevices with the dataplane.  This speeds up the initial sync (especially with large numbers of services).
```